+++ src/Core/Client/Adapter/TimeoutAwareInterface.php
@@ -22,6 +22,13 @@ interface TimeoutAwareInterface
     public const DEFAULT_TIMEOUT = 5;
 
     /**
+     * Fast timeout that should be used if the client should not wait for the result.
+     *
+     * @see \Solarium\Plugin\NoWaitForResponseRequest
+     */
+    public const FAST_TIMEOUT = 1;
+
+    /**
      * @param int $timeoutInSeconds
      *
      * @return self Provides fluent interface