+++ tests/Core/Client/Adapter/HttpTest.php
@@ -41,7 +41,7 @@ class HttpTest extends TestCase
         $mock->expects($this->once())
              ->method('getData')
              ->with($this->equalTo('http://127.0.0.1:8983/solr/'), $this->isType('resource'))
-             ->willReturn([$data, ['HTTP 1.1 200 OK']]);
+             ->willReturn([$data, ['HTTP/1.1 200 OK']]);
 
         $mock->execute($request, $endpoint);
     }