+++ tests/Core/Event/PreCreateResultTest.php
@@ -15,7 +15,7 @@ class PreCreateResultTest extends TestCase
         $client = TestClientFactory::createWithCurlAdapter();
         $query = $client->createSelect();
         $query->setQuery('test123');
-        $response = new Response('', ['HTTP 1.0 200 OK']);
+        $response = new Response('', ['HTTP/1.0 200 OK']);
 
         $event = new PreCreateResult($query, $response);
 
@@ -35,7 +35,7 @@ class PreCreateResultTest extends TestCase
         $client = TestClientFactory::createWithCurlAdapter();
         $query = $client->createSelect();
         $query->setQuery('test123');
-        $response = new Response('', ['HTTP 1.0 200 OK']);
+        $response = new Response('', ['HTTP/1.0 200 OK']);
 
         $result = new Result($query, $response);
         $event->setResult($result);