+++ tests/QueryType/Server/Api/ResponseParserTest.php
@@ -25,7 +25,7 @@ class ResponseParserTest extends TestCase
             }
         EOT;
 
-        $response = new Response($data, ['HTTP 1.1 200 OK']);
+        $response = new Response($data, ['HTTP/1.1 200 OK']);
         $result = new Result(new Query(), $response);
         $parser = new ResponseParser();
         $parsed = $parser->parse($result);