+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Option `buildAll` for Suggesters
+- NoWaitForResponseRequest Plugin
+
 ### Fixed
 - PHP 8.2 deprecations for Solarium\QueryType\Server\Collections results